+++ src/pydistcheck/_config.py
