+++ src/pydistcheck/_file_utils.py
