+++ src/pydistcheck/_inspect.py
@@ -4,7 +4,7 @@ Code that prints diagnostic information about a distribution.
 
 from typing import TYPE_CHECKING
 
-from .utils import _FileSize
+from ._utils import _FileSize
 
 if TYPE_CHECKING:
     from .distribution_summary import _DistributionSummary
