+++ src/pydistcheck/_utils.py
